refactor(element-template): split plugin file; type-safe subcommand dispatch
The plugin file was ~1500 lines covering six subcommands plus shared
helpers. Splitting it into sibling files lets c8ctl-plugin.ts cover
only CLI dispatch and the Plugin API, mirroring the bpmn plugin's
shape after lint.ts landed.

New files under default-plugins/element-template/:

- template-ref.ts — TemplateRef classification (url/path/id), template
  loading (readTemplateFromPathOrUrl, resolveOotbTemplate, loadTemplate),
  BPMN input reading and executionPlatformVersion extraction. Shared by
  apply, info, get, get-properties.
- apply.ts — applySubcommand plus vendor-bundle resolution and the
  apply-only dry-run renderer.
- info.ts — infoSubcommand, the shared parseInspectArgs (used by
  get-properties), the template-header renderer (used by search), the
  generic formatKeyedCard layout (used by get-properties' property
  cards), and the JSON summary projection (used by search).
- get-properties.ts — getPropertiesSubcommand and its renderers:
  condensed table, detail cards, group/glob filtering, JSON projection.
- get.ts — getSubcommand plus stripInternalMetadata / stripIcon.
- search.ts — searchSubcommand.
- sync.ts — syncSubcommand.

c8ctl-plugin.ts now derives the Subcommand union from the metadata
literal type and uses a Record<Subcommand, ...> dispatch table. Adding
a subcommand to metadata without a matching handler is now a TS error,
replacing the previous string-comparison if/else ladder.
